requirements.txt
torch==2.5.1+cu121 
torchvision==0.20.1+cu121
chromadb==1.3.5
sentence-transformers==5.1.2
transformers==4.57.3
accelerate==1.12.0
PyMuPDF==1.26.6
python-docx==1.2.0
pandas==2.3.3
pdfplumber==0.11.8
openpyxl==3.1.5
loguru==0.7.3

